HTOOLHANDLE is actually a pointer

+++ AfxHookSource/SourceInterfaces.h
@@ -5209,7 +5209,7 @@ enum
 //-----------------------------------------------------------------------------
 // Safe accessor to an entity
 //-----------------------------------------------------------------------------
-typedef unsigned int HTOOLHANDLE;
+typedef size_t HTOOLHANDLE;
 enum
 {
 	HTOOLHANDLE_INVALID = 0
